Dataset white and red wine
This is the dataset that I worked on in the project. It is related to white and red wines, each of which is characterized by 11 descriptive features and a quality feature that defines the class of the wine. 
The goal of the dataset analysis is to understand what are the main variables that describes each wine and how to use the machine learning algorithms to retrive the correct cluster